`NavigationCoordinator` moved to its own file.
